requirements.txt
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
astrapy==0.3.3
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
build==0.8.0
cachetools==5.2.0
cassandra-driver==3.25.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.7
click==8.0.3
cmarkgfm==2022.10.27
colorama==0.4.6
commonmark==0.9.1
conffu==2.2.23
configparser==5.3.0
cql==1.4.0
cqlsh==6.0.0
DateTime==5.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.5
django-cassandra-engine==1.7.0
docutils==0.19
fontawesome-markdown @ https://github.com/bmcorser/fontawesome-markdown/archive/master.zip
frozenlist==1.3.3
future==0.18.3
geomet==0.2.1.post1
ghp-import==2.1.0
gitdb==4.0.10
github==1.2.7
githubkit==0.10.3
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
gql==3.4.0
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.0
httpie==2.6.0
httpie-astra==0.1.0
httplib2==0.21.0
httpx==0.24.0
idna==3.3
importlib-metadata==5.2.0
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-git-revision-date-plugin==0.3.2
mkdocs-material==9.0.8
mkdocs-material-extensions==1.1.1
more-itertools==9.0.0
oauthlib==3.2.2
octohub==0.4
packaging==21.3
pandas==1.4.1
pep517==0.12.0
pkginfo==1.9.2
protobuf==4.21.10
py-gfm==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycmarkgfm==1.2.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
PyGithub==1.57
Pygments==2.14.0
PyJWT==2.6.0
pymdown-extensions==9.9.2
PyNaCl==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pytz==2022.1
PyYAML==6.0
pyyaml_env_tag==0.1
readme-renderer==37.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.6.0
rsa==4.9
scandir==1.10.0
scylla-driver==3.25.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
thrift==0.15.0
tomli==2.0.1
treewalker==0.1.9
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.7
watchdog==2.2.1
webencodings==0.5.1
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
zope.interface==5.5.2
